add keyhandling and RSA keygeneration
Currently stpm2 only is able to hold and handle a single key in the context.
A new key can be created using stpm2_create_rsa_2048(), calling this function
will automatically load the key into the TPM. If the user wants to create
a new key first the current key has the be unloaded by calling stpm2_unload_key(),
the key will also be unloaded from the TPM. The current key in the context
can also be exported in an openssl compatible format using stpm2_export_pubkey_pem().

Signed-off-by: Martin Pietryka <martin@pietryka.at>