examples: rename rsa -> rsa_keys
The example now generates a new RSA keypair and exports the .pem and the .bin
file to /tmp, then a new stpm2 session is started and the key is loaded again
and re-exported. Checking the keyfiles from the first and the second session
shows that they are the same, thus the export and import is most likely working.

Signed-off-by: Martin Pietryka <martin@pietryka.at>